requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Flask dependencies for web backend
Flask>=2.3.0
Flask-CORS>=4.0.0

# Optional: Voice support (uncomment if you want text-to-speech)
# pyttsx3>=2.90

# Optional: Enhanced voice recognition (for future features)
# vosk>=0.3.45
# pyaudio>=0.2.11

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0